rename variables and files to match endpoint path
